Suppress fake uninstall errors.
Check the first arg passed to the plugin, to make sure that it is actually `do-all`.
If you are uninstalling the plugin, it will still call the `Run()` method, but the
first argument will not be `do-all`. Simply return in this case to avoid error
messages in the uninstall output.